Fixed compiler error in HTMLKit when used in Objective-C project.
Because the header includes and inlined function definition, any project that references the HTMLKit headers would compile this function, which fails if the 'Sign Comparison' warning/error is enabled in the project's build settings.

Fixed the type of variable in the function.

+++ Sources/include/HTMLQuirksMode.h
@@ -84,7 +84,7 @@ static NSString * HTMLQuirksModePrefixes[] = {
 
 NS_INLINE BOOL QuirksModePrefixMatch(NSString *publicIdentifier)
 {
-	for (int i = 0; i < sizeof(HTMLQuirksModePrefixes) / sizeof(HTMLQuirksModePrefixes[0]); i++) {
+	for (unsigned long i = 0; i < sizeof(HTMLQuirksModePrefixes) / sizeof(HTMLQuirksModePrefixes[0]); i++) {
 		if ([publicIdentifier hasPrefixIgnoringCase:HTMLQuirksModePrefixes[i]]) {
 			return YES;
 		}
